+++ tests/test_data_bundle_processing.py
@@ -1,0 +1,16 @@
+from django.test import TestCase
+
+class TestDataBundleProcessing(TestCase):
+    def setUp(self):
+        pass
+
+    def test_todo(self):
+        # TODO: Test cleartext bundle
+        # TODO: Test encrypted bundle
+        # TODO: Test compressed bundle
+        # TODO: Test compressed+encrypted bundle
+        # TODO: Test bundle trace records
+        
+        # TODO: Test federation
+
+        pass
